[refactor] 1. IdGenerator Component 제거 2. static 메소드 사용 3. 리팩토링
