--- api-module/src/main/java/com/api/enums/IdPrefix.java
@@ -1,12 +1,0 @@
-package com.api.enums;
-
-public enum IdPrefix {
-
-    MDL,    /*model*/
-    PHO,    /*photographer*/
-    MSG,    /*message*/
-    CON,    /*conversation*/
-    MAT,    /*matching*/
-    FIL     /*file*/
-
-}
