+++ api-module/src/main/java/com/api/util/IdGenerator.java
@@ -1,22 +1,51 @@
 package com.api.util;
 
-import com.api.enums.IdPrefix;
-import org.springframework.stereotype.Component;
-
 /**
  * Id를 부여하는 Generator
  * id 채번 rule : {prefix}_UUID
  * prefix : IdPrefix
  *
  */
-@Component
 public class IdGenerator {
 
-    public String generateId(IdPrefix prefix) {
+    public static String generateId(IdPrefix prefix) {
         if (prefix == null) {
             throw new IllegalArgumentException("prefix is null");
         }
 
         return prefix + "_" + java.util.UUID.randomUUID().toString();
+    }
+
+    public static IdPrefix getPrefixModel() {
+        return IdPrefix.MDL;
+    }
+
+    public static IdPrefix getPrefixPhotographer() {
+        return IdPrefix.PHO;
+    }
+
+    public static IdPrefix getPrefixMessage() {
+        return IdPrefix.MSG;
+    }
+
+    public static IdPrefix getPrefixConversation() {
+        return IdPrefix.CON;
+    }
+
+    public static IdPrefix getPrefixMatching() {
+        return IdPrefix.MAT;
+    }
+
+    public static IdPrefix getPrefixFile() {
+        return IdPrefix.FIL;
+    }
+
+    public enum IdPrefix {
+        MDL,    /*model*/
+        PHO,    /*photographer*/
+        MSG,    /*message*/
+        CON,    /*conversation*/
+        MAT,    /*matching*/
+        FIL     /*file*/
     }
 }
